requirements.txt
astroid==2.4.2
autopep8==1.5.5
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.23
neo4j==4.2.1
protobuf==3.14.0
pycodestyle==2.6.0
pylint==2.6.2
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
toml==0.10.2
wrapt==1.12.1
